test: region > testes unitários de serviços
